use cet_make_library and cet_build_plugin; remove unused openmp instructions
